refactor(mirror): separate sync engine responsibilities
